move PlaceRating action, request, response, exception to convention format folders
