+++ app/Actions/Place/GetPlaceRatingAverage/GetPlaceRatingAvgRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Hedonist\Actions\Place\Rate;
+namespace Hedonist\Actions\Place\GetPlaceRatingAverage;
 
 class GetPlaceRatingAvgRequest
 {
