+++ app/Actions/Place/SetPlaceRating/SetPlaceRatingResponse.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Hedonist\Actions\Place\Rate;
+namespace Hedonist\Actions\Place\SetPlaceRating;
 
 class SetPlaceRatingResponse
 {
